backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn[standard]==0.32.0
python-multipart==0.0.20
psycopg[binary,pool]>=3.2.0
asyncpg>=0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0